Avoid a bogus gcc-15 warning for literal signedness.
It ignores explicitly given type, and then complains.

+++ tintin.h
@@ -234,7 +234,7 @@ typedef int64_t timens_t;
 typedef struct trip *ptrip;
 
 #define NANO 1000000000LL
-#define INVALID_TIME 0x8000000000000000LL
+#define INVALID_TIME ((int64_t)0x8000000000000000LL)
 
 #define ARRAYSZ(x) (sizeof(x)/sizeof((x)[0]))
